requirements.txt
requests==2.28.0
Flask==2.1.2
python-dotenv==0.20.0
google-cloud-storage==2.4.0